Split monolithic counter_culture_spec.rb into per-feature files
The 4,298-line spec/counter_culture_spec.rb is split into 22 focused
files under spec/counter_culture/, plus shared infrastructure in a new
spec/support/ directory (model requires, DatabaseCleaner setup,
yaml_load helper). The support glob in spec_helper.rb already auto-loads
files in that directory.

Each new file is self-contained (require 'spec_helper' + its own
RSpec.describe), follows the convention established by the existing
configuration_spec.rb, and can be run in isolation.

Test count unchanged: 211 examples pass (208 from the monolith +
3 from the existing configuration_spec.rb).
