+++ spec/support/database_cleaner.rb
@@ -1,0 +1,9 @@
+require 'database_cleaner'
+
+DatabaseCleaner.strategy = :deletion
+
+RSpec.configure do |config|
+  config.before(:each) do
+    DatabaseCleaner.clean
+  end
+end
