+++ spec/support/model_requires.rb
@@ -1,0 +1,56 @@
+require 'models/company'
+require 'models/comment'
+require 'models/industry'
+require 'models/product'
+require 'models/review'
+require 'models/simple_review'
+require 'models/twitter_review'
+require 'models/user'
+require 'models/admin_user'
+require 'models/category'
+require 'models/has_string_id'
+require 'models/has_non_pk_id'
+require 'models/simple_main'
+require 'models/simple_dependent'
+require 'models/conditional_main'
+require 'models/conditional_dependent'
+require 'models/post'
+require 'models/post_comment'
+require 'models/post_like'
+require 'models/categ'
+require 'models/soft_delete_discard'
+require 'models/soft_delete_paranoia'
+require 'models/subcateg'
+require 'models/another_post'
+require 'models/another_post_comment'
+require 'models/person'
+require 'models/transaction'
+require 'models/conversation'
+require 'models/candidate_profile'
+require 'models/candidate'
+require 'models/with_module/model1'
+require 'models/with_module/model2'
+require 'models/prefecture'
+require 'models/city'
+require 'models/group'
+require 'models/sub_group'
+require 'models/group_item'
+require 'models/article_group'
+require 'models/article'
+require 'models/sti_contract/agreement'
+require 'models/sti_contract/base'
+require 'models/locking_parent'
+require 'models/locking_child'
+require 'models/locking_touch_child'
+require 'models/locking_grandchild'
+
+if CounterCulture.supports_composite_keys?
+  require 'models/composite_group'
+  require 'models/composite_group_user'
+  require 'models/composite_user'
+end
+
+if ENV['DB'] == 'postgresql'
+  require 'models/purchase_order'
+  require 'models/purchase_order_item'
+end
