+++ spec/support/yaml_helper.rb
@@ -1,0 +1,9 @@
+module YamlHelper
+  def yaml_load(yaml)
+    YAML.safe_load(yaml, permitted_classes: [Time])
+  end
+end
+
+RSpec.configure do |config|
+  config.include YamlHelper
+end
